requirements.txt
discord.py
requests
python-decouple
mysql-connector-python